Move space subscription hook related logic
