+++ components/settings/subscription/SubscriptionSettings.tsx
@@ -4,17 +4,16 @@ import { Divider } from '@mui/material';
 import { Stack } from '@mui/system';
 import { Elements } from '@stripe/react-stripe-js';
 import { useEffect, useState } from 'react';
-import useSWR from 'swr';
 import useSWRMutation from 'swr/mutation';
 
 import charmClient from 'charmClient';
 import { useSnackbar } from 'hooks/useSnackbar';
 import { useSpaces } from 'hooks/useSpaces';
-import { useWebSocketClient } from 'hooks/useWebSocketClient';
 import type { UpdateSubscriptionRequest } from 'lib/subscription/interfaces';
 import { inputBackground } from 'theme/colors';
 
 import { CheckoutForm } from './CheckoutForm';
+import { useSpaceSubscription } from './hooks/useSpaceSubscription';
 import { loadStripe } from './loadStripe';
 import { SubscriptionActions } from './SubscriptionActions';
 import { SubscriptionInformation } from './SubscriptionInformation';
@@ -22,16 +21,8 @@ import { SubscriptionInformation } from './SubscriptionInformation';
 export function SubscriptionSettings({ space }: { space: Space }) {
   const { showMessage } = useSnackbar();
   const { setSpace } = useSpaces();
-  const { subscribe } = useWebSocketClient();
 
-  const {
-    data: spaceSubscription,
-    isLoading,
-    mutate: refetchSpaceSubscription
-  } = useSWR(`${space.id}-subscription`, () => charmClient.subscription.getSpaceSubscription({ spaceId: space.id }), {
-    shouldRetryOnError: false,
-    revalidateOnFocus: false
-  });
+  const { spaceSubscription, isLoading, refetchSpaceSubscription } = useSpaceSubscription();
 
   const { trigger: updateSpaceSubscription, isMutating: isLoadingUpdate } = useSWRMutation(
     `/api/spaces/${space?.id}/subscription`,
@@ -43,16 +34,6 @@ export function SubscriptionSettings({ space }: { space: Space }) {
       }
     }
   );
-
-  useEffect(() => {
-    const unsubscribeFromSpaceSubscriptionUpdates = subscribe('space_subscription', () => {
-      refetchSpaceSubscription();
-    });
-
-    return () => {
-      unsubscribeFromSpaceSubscriptionUpdates();
-    };
-  }, []);
 
   const { trigger: deleteSubscription, isMutating: isLoadingDeletion } = useSWRMutation(
     `/api/spaces/${space?.id}/subscription-intent`,
